refactor(krites/query): full structural overhaul — snafu errors, quality (#3215)
Split monolithic ra/sources.rs (793 LOC) into focused modules:
- ra/inline_fixed.rs: constant data sources
- ra/stored.rs: persistent relation scan (with/without time-travel)
- ra/temp_store.rs: derived relation sources for semi-naive eval

Added comprehensive documentation throughout:
- Pipeline overview in query/mod.rs (logical -> reorder -> stratify -> magic -> compile -> eval)
- Magic sets algorithm reference (Beeri & Ramakrishnan 1991)
- Stratification rules (Tarjan SCC, generalized Kahn, meet-aggregation monotonicity)
- Graph algorithm descriptions (Tarjan SCC, reachability, generalized Kahn)
- Every RA node type documented with purpose and complexity
- Join strategy selection documented (prefix vs materialized vs point-lookup)

Extracted TempStoreRA.resolve_store() helper to DRY epoch store lookup.
Removed duplicate BindingFormatter from join.rs (was in both mod.rs and join.rs).
Tightened lint expectations to match actual usage per split module.
